fix naming mistakes
