+++ src/main/java/com/ostapdev/todo/parser/CommandParser.java
@@ -1,16 +1,16 @@
 package com.ostapdev.todo.parser;
 
-import com.ostapdev.todo.parser.command.Command;
+import com.ostapdev.todo.parser.command.BaseCommand;
 import com.ostapdev.todo.parser.command.CommandFactory;
 
 import java.util.List;
 
 public class CommandParser {
-    private final List<Command> commands = CommandFactory.getCommands();
+    private final List<BaseCommand> commands = CommandFactory.getCommands();
 
     public void parse(String inputLine){
-        for (Command command : commands){
-            command.checkInput(inputLine);
+        for (BaseCommand command : commands){
+            command.run(inputLine);
         }
     }
 }
