+++ src/main/java/com/ostapdev/todo/parser/command/CommandFactory.java
@@ -3,7 +3,7 @@ package com.ostapdev.todo.parser.command;
 import java.util.List;
 
 public class CommandFactory {
-    public static List<Command> getCommands(){
+    public static List<BaseCommand> getCommands(){
         return List.of(new AddCommand(),new PrintCommand(),new DeleteCommand(),new EditCommand(),new QuitCommand(),
                 new SearchCommand(),new ToggleCommand());
     }
