--- src/main/java/com/ostapdev/todo/parser/command/CommandInterface.java
@@ -1,5 +1,0 @@
-package com.ostapdev.todo.parser.command;
-
-public interface CommandInterface {
-    void checkInput(String inputLine);
-}
